Fix for prevent tool reported defects.
Change details:
1. Prevent issues fix.
2. Removal of white spaces from Java files.
3. Simulator tool UI changes.
4. Fix for issues observed during regression testing.

Reviewed-on: https://gerrit.iotivity.org/gerrit/2593
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>

+++ service/simulator/java/jni/resource_attributes_jni.cpp
@@ -165,7 +165,7 @@ bool JResourceAttributeConverter::setType(JNIEnv *env, jobject &jattributeObj)
         return false;
     }
 
-    int type = m_attribute.getValueType();
+    int type = static_cast<int>(m_attribute.getValueType());
     jobject jtype = env->CallStaticObjectMethod(clsType, midGetType, type);
     if (env->ExceptionCheck())
     {
